refactor: move Fetcher
